add rate and add review

+++ BoardGameApp/src/main/java/org/openjfx/DBManager/Neo4jDBManager/GamesReviewsRatesDBManager.java
@@ -128,7 +128,7 @@ public class GamesReviewsRatesDBManager {
         return numberRates;
     }
 
-    public static int avgRates(String name)
+    public static float avgRates(String name)
     {
         try(Session session=driver.session())
         {
